Move TypeNameHelper to independant file.
