+++ src/matrix/operator.cpp
@@ -3,6 +3,7 @@
  * François Bleibel,
  * Olivier Stasse,
  * Nicolas Mansard
+ * Joseph Mirabel
  *
  * CNRS/AIST
  *
@@ -35,6 +36,8 @@
 #include <boost/numeric/conversion/cast.hpp>
 #include <deque>
 
+#include "../tools/type-name-helper.hh"
+
 namespace dg = ::dynamicgraph;
 
 /* ---------------------------------------------------------------------------*/
@@ -43,27 +46,6 @@ namespace dg = ::dynamicgraph;
 
 namespace dynamicgraph {
   namespace sot {
-    template< typename TypeRef >
-    struct TypeNameHelper
-    {
-      static const std::string typeName;
-    };
-    template< typename TypeRef >
-    const std::string TypeNameHelper<TypeRef>::typeName = "unspecified";
-
-#define ADD_KNOWN_TYPE( typeid ) \
-    template<>const std::string TypeNameHelper<typeid>::typeName = #typeid
-
-    ADD_KNOWN_TYPE(bool);
-    ADD_KNOWN_TYPE(double);
-    ADD_KNOWN_TYPE(Vector);
-    ADD_KNOWN_TYPE(Matrix);
-    ADD_KNOWN_TYPE(MatrixRotation);
-    ADD_KNOWN_TYPE(MatrixTwist);
-    ADD_KNOWN_TYPE(MatrixHomogeneous);
-    ADD_KNOWN_TYPE(VectorQuaternion);
-    ADD_KNOWN_TYPE(VectorRollPitchYaw);
-
     template< typename TypeIn, typename TypeOut >
     struct UnaryOpHeader
     {
@@ -81,8 +63,6 @@ namespace dynamicgraph {
 	  std::string ("\n");
       }
     };
-
-
   } /* namespace sot */
 } /* namespace dynamicgraph */
